2023_09_05(use silhouette score for cluster)

+++ bin/main.py
@@ -639,16 +639,18 @@ def main(input_file, genome_file, output_dir, continue_analysis, pfam_dir, min_b
 
     final_con_file = os.path.join(output_dir, "TE_Trimmer_consensus.fasta")
 
-    # Read the file into memory
-    with open(final_con_file, 'r') as file:
-        file_contents = file.read()
+    if os.path.exists(final_con_file):
 
-    # Make the replacements
-    file_contents = file_contents.replace('___', '#').replace('__', '/')
+        # Read the file into memory
+        with open(final_con_file, 'r') as file:
+            file_contents = file.read()
 
-    # Write the modified contents back to the file
-    with open(final_con_file, 'w') as file:
-        file.write(file_contents)
+        # Make the replacements
+        file_contents = file_contents.replace('___', '#').replace('__', '/')
+
+        # Write the modified contents back to the file
+        with open(final_con_file, 'w') as file:
+            file.write(file_contents)
 
     # At the end of the program, check if all sequences have been processed
     with open(progress_file, 'r') as file:
